Removes connected_port, implements connected_ports
This work starts the effort of removing connected_port from the specs.

I have moved the code around a bit to get all connected_ports, the tests seem
to pass in the appropriate areas.

I want to verify this path before attempting to remove all UniquePort calls in
the tests.